Refactor output layout
